% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SgsLoad.R
\name{as_Matrix}
\alias{as_Matrix}
\title{Deal large matrix}
\usage{
as_Matrix(mat)
}
\arguments{
\item{mat}{a big dgMatrix}
}
\value{
Return a matrix
}
\description{
Deal large matrix
}
\examples{
\dontrun{
counts <- as_Matrix(counts)
}


}
