% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SgsLoad.R
\name{ExportSC}
\alias{ExportSC}
\title{Export the seurat object into SGS}
\usage{
ExportSC(
  object,
  species_id,
  track_name,
  track_type,
  select_group,
  assays,
  matrix.slot,
  assay.type,
<<<<<<< HEAD
=======
  dir_path,
>>>>>>> 943979c5ecf8cbec5baf2ebbbca37c6b53c9225e
  reductions = NULL,
  marker.files = NULL
)
}
\arguments{
\item{object}{seurat object used to load}

\item{species_id}{the id of the species}

\item{track_name}{the name of the track}

\item{track_type}{type of the track,mainly:"sc_RNA"; "sc_ATAC"}

\item{select_group}{a vector of cell group name to export}

\item{assays}{a vector of assay names to export: c("RNA","ATAC","motif")}

\item{matrix.slot}{a list of matrix data to export: list("RNA"="data", "ATAC"="counts"; "motif"="data")}

\item{assay.type}{a list of matrix type, mainly:"peak", "motif", "gene".list("RNA"="gene", "ATAC"="peak"; "motif"="motif")}

<<<<<<< HEAD
=======
\item{dir_path}{the output dir}

>>>>>>> 943979c5ecf8cbec5baf2ebbbca37c6b53c9225e
\item{reductions}{a vector of reduction names to export}

\item{marker.files}{a list of marker file path to export.list("RNA"="xxx.rna.tsv", "ATAC"="xxx.peak.tsv"; "motif"="xxx.motif.tsv")}
}
\value{
a post information in json in format
}
\description{
Export the seurat object into SGS
}
\examples{
\dontrun{
#### Loadding the scRNA-seq object into SGS
<<<<<<< HEAD
library(Signac)
=======
>>>>>>> 943979c5ecf8cbec5baf2ebbbca37c6b53c9225e
data("pbmc_small")
marker_file <- system.file("extdata/scRNA", "markers.tsv", package = "sgsload")

result_json <- ExportSC(
  object = pbmc_small,
<<<<<<< HEAD
  species_id = "678d0bb42ebd4137b031ec5cc90dc0c5",
=======
  species_id = "00000",
>>>>>>> 943979c5ecf8cbec5baf2ebbbca37c6b53c9225e
  track_name = "pbmc_small",
  track_type = "scRNA",
  select_group = c("groups", "RNA_snn_res.0.8", "letter.idents", "cluster"),
  assays = c("RNA"),
  matrix.slot = list("RNA" = "data"),
  assay.type = list("RNA" = "gene"),
<<<<<<< HEAD
=======
  dir_path = "./out",
>>>>>>> 943979c5ecf8cbec5baf2ebbbca37c6b53c9225e
  reductions = c("tsne", "umap"),
  marker.files = list("RNA" = marker_file)
)


#### Loadding the scATAC-seq object into SGS
data("atac")
marker_genes <- system.file("extdata/scATAC", "marker_genes.tsv", package = "sgsload")
marker_motifs <- system.file("extdata/scATAC", "marker_motif.tsv", package = "sgsload")
marker_peaks <- system.file("extdata/scATAC", "marker_peaks.tsv", package = "sgsload")
markers <- list(marker_genes, marker_motifs, marker_peaks)
names(markers) <- c("RNA", "chromvar", "peaks")

# change the fragment path
new.path <- system.file("extdata/scATAC", "fragments.tsv.gz", package = "sgsload")
fragments <- CreateFragmentObject(
  path = new.path,
  cells = colnames(atac),
  validate.fragments = TRUE
)
Fragments(atac) <- NULL
Fragments(atac) <- fragments


result_json <- ExportSC(
  object = atac,
<<<<<<< HEAD
  species_id = "678d0bb42ebd4137b031ec5cc90dc0c5",
=======
  species_id = "00000",
>>>>>>> 943979c5ecf8cbec5baf2ebbbca37c6b53c9225e
  track_name = "atac",
  track_type = "scATAC",
  select_group = c("seurat_clusters"),
  assays = c("RNA", "chromvar", "peaks"),
  matrix.slot = list("RNA" = "data", "chromvar" = "data", "peaks" = "data"),
  assay.type = list("RNA" = "gene", "chromvar" = "motif", "peaks" = "peak"),
<<<<<<< HEAD
=======
  dir_path = "./out",
>>>>>>> 943979c5ecf8cbec5baf2ebbbca37c6b53c9225e
  reductions = c("lsi", "umap"),
  marker.files = markers
)
}

}
\author{
xtt
}
