% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SgsLoad.R
\name{get_species_inform}
\alias{get_species_inform}
\title{Get the species id information}
\usage{
get_species_inform(user_id = user_id)
}
\arguments{
\item{user_id}{a string of the user id, defualt user id is "user001".}
}
\value{
a json string of the species id information
}
\description{
Get the species id information
}
\examples{
\dontrun{
get_species_inform(user_id)
}


}
\author{
xtt
}
