% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_small}
\alias{pbmc_small}
\title{A small example scRNA-seq dataset
A subsetted version of 10x Genomics 10x human(hg19) PBMC scRNA-seq dataset}
\format{
A Seurat object with the following assays
\describe{
\item{RNA}{A gene x cell dataset}
}
}
\usage{
pbmc_small
}
\description{
A small example scRNA-seq dataset
A subsetted version of 10x Genomics 10x human(hg19) PBMC scRNA-seq dataset
}
\concept{data}
\keyword{datasets}
